Added proper version checking and updating
The app is now able to update itself so you no longer have to manually update the application yourself.

This works for both if you are running the .py file and when your are running the .exe, depending on which you are running it will update the correct file.

We now also show a screen displaying what has changed in the new version since the version you are running.